AttemptFour/requirements_train.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
apptools==5.1.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
black==22.1.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer==2.0.12
cifti==1.1
click==8.0.3
conda==4.11.0
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262148928/work
configobj==5.0.6
cryptography @ file:///tmp/build/80754af9/cryptography_1639400846433/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython==0.29.27
decorator==5.1.1
envisage==6.0.1
executing==0.8.2
flatbuffers==2.0
fonttools==4.25.0
frozenlist==1.3.0
future==0.18.2
gast==0.3.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==2.10.0
html5lib==1.1
idna==3.3
imageio==2.15.0
importlib-metadata==4.11.0
importlib-resources==5.4.0
ipython==8.0.1
jedi==0.18.1
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282420641/work
libclang==13.0.0
lxml==4.7.1
Markdown==3.3.6
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1638289681807/work
matplotlib-inline==0.1.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
multidict==6.0.2
munkres==1.1.4
mypy-extensions==0.4.3
networkx==2.6.3
nibabel==3.2.2
nilearn==0.9.0
nltk==3.7
nsd-access @ file:///home/seagie/nsd_access
numexpr @ file:///tmp/build/80754af9/numexpr_1640704208950/work
numpy==1.18.5
oauthlib==3.2.0
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.5
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.0
prompt-toolkit==3.0.28
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycortex==1.2.3
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyface==7.4.0
Pygments==2.11.2
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016699510/work
scikit-learn==1.0.2
scipy==1.4.1
Shapely==1.8.0
sip==4.19.13
six==1.16.0
sklearn==0.0
stack-data==0.1.4
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.0
tensorflow-addons==0.15.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tomli==2.0.1
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm @ file:///tmp/build/80754af9/tqdm_1635330843403/work
traitlets==5.1.1
traits==6.3.2
traitsui==7.2.1
typeguard==2.13.3
typing_extensions==4.0.1
urllib3==1.26.8
vtk==9.1.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wget==3.2
wrapt==1.13.3
wslink==1.4.1
yarl==1.7.2
zipp==3.7.0
